+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+
+## [0.2.0] - 2026-08-27
+
 ### Added
 - `sicd_alias` CLI subcommand
 
@@ -26,5 +29,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `remocomp` module
 - `sicd_pixel_type` module
 
-[unreleased]: https://github.com/ValkyrieSystems/sarkit-processing/compare/v0.1.0...HEAD
+[unreleased]: https://github.com/ValkyrieSystems/sarkit-processing/compare/v0.2.0...HEAD
+[0.2.0]: https://github.com/ValkyrieSystems/sarkit-processing/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/ValkyrieSystems/sarkit-processing/releases/tag/v0.1.0